consolidate user invite use cases and actions
